refactor: extract shared presentation support and version metadata
